requirements.txt
arrow==0.17.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
cassiopeia==4.0.14
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
datapipelines==1.0.7
decorator==4.4.2
dnspython==2.0.0
fake-useragent==0.1.11
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
merakicommons==1.0.9
moviepy==1.0.3
networkx==2.5
numpy==1.19.2
oauth2client==4.1.3
Pillow==7.2.0
proglog==0.1.9
progressbar2==3.53.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.0
pyser==0.1.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-utils==2.4.0
pytz==2020.1
requests==2.24.0
requests-file==1.5.1
rsa==4.6
selenium==3.141.0
selenium-firefox==0.5.18
simple-youtube-api==0.2.8
six==1.15.0
soupsieve==2.0.1
tldextract==2.2.3
tqdm==4.50.2
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
youtube-uploader-selenium==0.1.0
